third_party/blink/web_tests/external/wpt/fetch/api/redirect/redirect-method.any-expected.txt

This is a testharness.js-based test.
PASS Response.redirected should be false on not-redirected responses
PASS Redirect 301 with GET
FAIL Redirect 301 with POST assert_equals: Request Content-Encoding after redirection is NO expected "NO" but got "Identity"
PASS Redirect 301 with HEAD
PASS Redirect 302 with GET
FAIL Redirect 302 with POST assert_equals: Request Content-Encoding after redirection is NO expected "NO" but got "Identity"
PASS Redirect 302 with HEAD
PASS Redirect 303 with GET
FAIL Redirect 303 with POST assert_equals: Request Content-Encoding after redirection is NO expected "NO" but got "Identity"
PASS Redirect 303 with HEAD
FAIL Redirect 303 with TESTING assert_equals: Request Content-Encoding after redirection is NO expected "NO" but got "Identity"
PASS Redirect 307 with GET
PASS Redirect 307 with POST (string body)
PASS Redirect 307 with POST (blob body)
PASS Redirect 307 with HEAD
Harness: the test ran to completion.
